+++ example/peripheral/sdio/sdkconfig.h
@@ -150,6 +150,11 @@
 
 #define CONFIG_FREERTOS_USE_FSDIO
 /* end of Freertos MMC Drivers */
+
+/* Freertos Adc Drivers */
+
+/* CONFIG_FREERTOS_USE_ADC is not set */
+/* end of Freertos Adc Drivers */
 /* end of Component Configuration */
 
 /* FreeRTOS Setting */